+++ CHANGES.markdown
@@ -2,6 +2,7 @@
  - `Functor` instances for `Handler`
  - `Functor`. `Applicative`, `Alternative`, and `MonadPlus` instances for
    `ArrowMonad`
+ - Expose `Read` and `Show` instances for `Down` on GHCs before 7.8
 
 ## Changes in 0.4.1
  - Fixed imports on GHC < 7.8 on Windows