+++ CHANGES.markdown
@@ -1,4 +1,4 @@
-## Changes in next
+## Changes in 0.4.2
  - `Functor` instances for `Handler`
  - `Functor`. `Applicative`, `Alternative`, and `MonadPlus` instances for
    `ArrowMonad`